Start a test suite
Right now we just have one test, which compares the example that we used
to have in a text file to the expected Ast.

A few things needed to be fixed modified in the source proper; had some
build errors that I was surprised by, and I needed an Eq instance for
asts.